Renamed/Fixed Names + Conditions + Connections
Changed two names in East Storage room to better reflect how the rest of the locations are named throughout the apworld, and removed condition on the heart door in that same room since the region connection has the condition. Did the same in the Parking Garage for the Elevator Control Room condition.

Flipped Side Stairs/Chief's Office Connection for those that use UT (maybe affected logic, too), since it was saying they had access to the Chief's Office without either condition the way it was before. Removed "Forbid: Square Crank" from some locations on Claire's side of the game since she doesn't ever receive it, either.

Changed Supply Room Weapon Cage Condition to require King Plug, since it's needed to access that location (thought we had done this before)

Changed Condition in Clock Tower to require Medallions instead of Electronic Boxed Part 2, so as to better indicate when the Clock Tower is available to the player if they use UT.

And finally, removed "Spade Key" as a requirement for conditions for both characters in B scenario, since it is now randomized into the pool.